WIP: Remove OpenApi objects with TOP extension
We need to remove the Swagger OpenAPI containers that have
the TOP extensions. These extensions were added as a workaround for
the ODL behavior where the PUT and POST requests should be executed
on different paths for the same object.

This workaround also resolves problems with the conversion of JSON to XML
by OpenAPI. OpenAPI creates an XML with a top element that
JSON does not require.

Currently, there is an issue with the PUT request and JSON, as well as
the POST request and XML. These issues need to be resolved.

JIRA: NETCONF-938
Change-Id: I68c4785299fda16480df8092a499ea42adc130c6
Signed-off-by: Peter Suna <peter.suna@pantheon.tech>