Rehost netconf-monitoring operations
The individual operations are not dependent on anything MD-SAL, rehost
them into the baseline netconf-server. The only remaining thing in
mdsal-netconf-monitoring is MonitoringToMdsalWriter, whose final
destination is still to be determined.

This is an intermediate step, as we are still looking for the set of
services provided at the plugins/ level, so a further movement is
likely.

JIRA: NETCONF-945
Change-Id: Ib77bea74ff242b22a21cbb9f0f0b8cb7990dee5a
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>