Improve NETCONF session ID handling
The session ID has a well-defined semantics, which are not a String, nor
a long, but rather an Uint32 with minimum value of 1 -- as captured by
SessionIdType.

Use SessionIdType as the proper type-safe capture of this construct
across the project.

This also forces netconf-testtool to use proper lifecycle hooks, so that
the session ID is propagated.

JIRA: NETCONF-945
Change-Id: If197be95e8ac4943409eb4e00476a57269828e18
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>