Change hardcoded actions path and provide tests
There was an assumption about the basePath ('rests') when creating
path for POST method for actions. The path was incorrect if the default
basePath was not used.

This patch fixes the path creation while not making any
concrete assumptions about the basePath. The only assumption
is that the basePath does NOT include FORWARD_SLASH character: '/'

JIRA: NETCONF-1021
Change-Id: I10d02cc13562ab815815e5cc1854309e030f8c56
Signed-off-by: Šimon Ukuš <simon.ukus@pantheon.tech>