Improve DefinitionGenerator testing
This commit addresses the lack of test coverage for various sections
within the DefinitionGenerator class. By adding more tests, we
significantly enhance the overall coverage of this class.
This improvement ensures a more robust and reliable testing environment.

The existing test also had to be modified to match the latest patch
where the default keyword was misused, and was replaced by example.

JIRA: NETCONF-999
Signed-off-by: tobias.pobocik <tobias.pobocik@pantheon.tech>
Change-Id: Iac703199c62f56b40dc02ff992e001a7b56554bf