Refactor test to verify the dynamicity of basePath
The previous change related to NETCONF-1021 modified
the nature of basePath. It is no longer hard coded as rests,
but can be specified by the user.

This change adjusts test to be parameterized, to verify the
correctness of the previous change by using the default basePath
along with a custom one.

JIRA: NETCONF-1021
Change-Id: I6c5bf49f2c2e0b34a37636886c2848ac904ca0a3
Signed-off-by: Šimon Ukuš <simon.ukus@pantheon.tech>