Fix incorrect union type assignment
In this commit, the issue with the union type assignment
for string, boolean and number has been addressed. Previously,
when using the union type, the string value was being
registered instead of the expected boolean or number value.

The necessary changes have been made to ensure that the
union type correctly recognizes and handles values.
Now, when working with this union type, the program will
properly identify and differentiate between, allowing for
accurate processing and manipulation of data.

JIRA: NETCONF-1058
Signed-off-by: tobias.pobocik <tobias.pobocik@pantheon.tech>
Change-Id: I75c2e42fbfa2a5562c01a8303828826adeeae600