Bump upstreams
Adopt:
- odlparent-13.0.1
- infrautils-6.0.0
- yangtools-11.0.0-SNAPSHOT
- mdsal-12.0.0-SNAPSHOT
- controller-8.0.0-SNAPSHOT
- aaa-0.18.0-SNAPSHOT

Handling of AugmentationIdentifier/AugmentationNode has been removed,
as this indirection is no longer present in yang-data-api.

The handling of PathArguments for the purposes of
{Netconf,Writer}FieldsTranslator has been updated, as yangtools no
longer provide wildcard PathArguments.

Also the handling of DTCLs is updated, so that we serializer changed
nodes in the same namespace first and then proceed to emit other
namespaces in lexical order.

JIRA: NETCONF-841
Change-Id: I6e8b9e834a33711387495e2cae52371ac456553d
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>