Correct YangInstanceIdentifierSerializerTest
Addressing modes for 'list' statements differs when they have a 'key'
statement from when they do not: former takes
NodeIdentifierWithPredicates to identify children, which the latter
takes NodeIdentifiers.

Expand the test cases to differentiate between the two -- note that
upstream accepts NodeIdentifierWithPredicates even for unkeyed lists,
but that is about to change.

Change-Id: Ib13d43097f86a6932e63e33d3a48788abe7640ec
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>