Add Child Nodes Only query parameter to SSE events
For SSE or Websocket, events are reported from the root node of the listen
Data Tree or from the modified leafs with the 'odl-leaf-nodes-only=true'
query parameter. This patch introduce a new 'odl-child-nodes-only' query
parameter to get only modified children in a compact form. This new option
could be considered as something in the middle between getting the whole
Data Tree from the root node and a per leaf node report.

JIRA: NETCONF-1076
Signed-off-by: Olivier Dugeon <olivier.dugeon@orange.com>
Change-Id: Ibeda878a38a52f911dd8fbaa0503c4bfb56ea9b7