Change schema for GET XML example
The XML example for GET request in OpenApi UI was different
to the one being actually returned. This patch changes the schema
used for that example, making it consistent with what is actually
returned in the response.

We still keep support for OAversion.V2_0, for which the change was
not applied. That is because this version does not support different
schemas for different MIME types

JIRA: NETCONF-1059
Change-Id: Ie2d128b6de43d705c93d509c287cabcbc719e591
Signed-off-by: Šimon Ukuš <simon.ukus@pantheon.tech>
Signed-off-by: Ivan Hrasko <ivan.hrasko@pantheon.tech>
(cherry picked from commit af6b9fe02a26fdd789470f0e8ab4c9062202c3a9)