OpenaApi: Use empty contet for operational GET
Previously, operational data was included in schemas
to provide a GET example for operational containers and lists.
However, removing this operational data caused an error in the OpenAPI UI
since the expected data was not found. To address this issue,
the operational data can be replaced with empty content instead.
Creating a duplicate schema specifically for this use case is both
memory-intensive and unnecessary.

This commit solves the problem by adding empty example
for the operational GET data.

JIRA: NETCONF-1061
Change-Id: I1444aad588c42417e5ff21dee46ff91ec0a1a07d
Signed-off-by: Peter Suna <peter.suna@pantheon.tech>