Remove operational GET request from OpenApi
Previously, operational data was included in schemas
to provide a GET example for operational containers and lists.
However, removing this operational data caused an error in the OpenAPI UI
since the expected data was not found. To address this issue,
the operational data can be replaced with empty content instead.
Creating a duplicate schema specifically for this use case is both
memory-intensive and unnecessary.

This commit solves the problem by completly removing Operational
GET request from OpenApi.

JIRA: NETCONF-1061
Change-Id: Ie6703a161210a27734e4cefdc254a1e0a3c858eb
Signed-off-by: Peter Suna <peter.suna@pantheon.tech>