Do not change NetconfSessionPromise to failure
Setting NetconfSessionPromise to failure when session negotiation fails
prevents recovery from a temporary session negotiation issue.
The state change does not prevent a new connection attempt but the
NETCONF session can never be fully established successfully even if the
negotiation itself is successful because it is tried with the same
promise which has already been set to failure.

Do not change the promise to failure when the session negotiation fails,
and leave the decision to ReconnectStrategy.

JIRA: NETCONF-1067
Change-Id: Id021d268f80f9991a2517a20ed105d3c1d87069c
Signed-off-by: Sangwook Ha <sangwook.ha@verizon.com>
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>