Rework NETCONF client reconnection
This patch reworks the way establishing a connection works: we only have
createClient(), which returns a simple future. The logic for
reconnecting the client is now owned by each individual app.

This affects only netconf-topology implementations, which share common
code here in NetconfConnectorDTO and AbstractNetconfTopology. This is
re-shuffled into NetconfNodeHandler, which now encapsulates all the
state as well as the logic to issue reconnects.

JIRA: NETCONF-1070
Change-Id: I8f175329ab2826e46c95dee59c408e71bf453ccf
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>