Split out NetconfClientConfigurationBuilderFactory
The task of creating a client configuration is independent of the actual
topology implementation. Rather that requiring a full-blown
AbstractNetconfTopology to accomplish this, split it out into a
self-contained component.

This component is then injected into topologies instead of its
constituent services -- which paves the way to expose NetconfNodeHandler
at a later point in time.

This also allows us to move a few tests from netconf-topology-impl to
netconf-topology-spi.

JIRA: NETCONF-1039
Change-Id: I9b92b662d6daf3106c26cbc6e470325bc5153348
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>