Use simple Executor in NetconfDevice
Passing down ListenableExecutorService ends up being interesting, as it
is not clear what the actual lifecycle is. Luckily we can get by using a
simple Executor with Futures.submit(), which essentially performs the
same task.

This makes it clear that the service is not touched by anyone in the
stack and we can use any old Executor -- to the point of using
directExecutor() in NetconfDeviceTest.

Change-Id: I31fcfcb7b303046935392bffd6d50c8f83ec33a2
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>