Fix device data update failure after successful connection
NetconfNodeHandler does not propagate successful connection to its
RemoteDeviceHandler. Make sure that it's notified so that the device
data is updated accordingly.

And call setFailure() in NetconfSessionPromise only if the promise has
not been set yet to avoid raising an unnecessary exception by trying
setFailure() a second time.

JIRA: NETCONF-1074
Change-Id: Id32996c11d0a63812d439ad386472c059255cf4e
Signed-off-by: Sangwook Ha <sangwook.ha@verizon.com>