Convert securitySchema's basicAuth field to record
This change replaces the JsonNode for a custom Java record that
contains only fields relevant to the Basic HTTP authentication scheme.

It introduces enum Type for Security Scheme Object type, which for now
contains only http, since that is the only type used in this project.
Another novelty is SecurityObject record that is, again, tailored
to be used with http type. Fields used with other types are not
included for the same reason the Type enum contains just http value.

JIRA: NETCONF-1075
Change-Id: I2a53b9002871e19808eef95739a07f5acd958815
Signed-off-by: Šimon Ukuš <simon.ukus@pantheon.tech>