Allow multiple security entries in securitySchema
The swagger docs specifies that the securitySchema is a map of entries
with the key being a custom name of the security method and the value
being an instance of a Security Scheme Object.

This change replaces the use of JsonNode etc., which added extra
overhead, for classes from java API and custom records.
It reworks the securitySchemes field in Components to be a map
and introduces an interface SecuritySchemeObject to be implemented
by records that represent the specific security object type.

It also introduces an Http record - an implementation of said interface.
Others might come in the future.

JIRA: NETCONF-1075
Change-Id: I2a53b9002871e19808eef95739a07f5acd958815
Signed-off-by: Šimon Ukuš <simon.ukus@pantheon.tech>
Signed-off-by: Ivan Hrasko <ivan.hrasko@pantheon.tech>