Remove POST requests for child elements
Currently POST requests have invalid examples. Creating requests for
child elements will cause some issues that not solvable with current
logic.

The purpose of this patch is to find out if we can left POST for at
least 1 elemend of the module or all POST should be removed.

JIRA: NETCONF-1071
Change-Id: I3762639dd4a4899ad431ef8109250a8191f82b2b
Signed-off-by: Oleksandr Zharov <oleksandr.zharov@pantheon.tech>