Make deviceName not to be null
We are passing null value of deviceName to methods only if the request,
for swagger document, is for controller (not a mounted device).
In that case we want "Controller" in place of the deviceName.

This change removes @Nullable annotations around String deviceName
and reworks the calls to methods, to pass "Controller" instead of
null value.

JIRA: NETCONF-1089
Change-Id: I5b37419f81c12c21f4ef65cc90e7b8dceb79995c
Signed-off-by: Šimon Ukuš <simon.ukus@pantheon.tech>
Signed-off-by: Ivan Hrasko <ivan.hrasko@pantheon.tech>