Rework tags creation in OpenApi
The tags values inside Operation object, according to Swagger spec
https://swagger.io/specification/#operation-object, do not have
any specific requirement. We are constructing them based on module's
origin (controller/mounted device) and module name,
storing it in an ArrayNode.

This change replaces the ArrayNode for List<String> and reworks
the contents by splitting the single tag into 2: one is the origin
the other is the module name.

JIRA: NETCONF-1087
Change-Id: I3a75b57f885503780867702a400d1be1c3ec6567
Signed-off-by: Šimon Ukuš <simon.ukus@pantheon.tech>