Remove addOperations()
This is a static utility method which builds an object and populates it
to a Map -- but there are only two callers and what gets put and where
and under which name is quite conflated.

Introduce a buildPostPath(), which just returns a Path and use that
instead.

While we are at it, use List.of() to specify empty inherited parameters.

Change-Id: I7d69af0ba96caecae3f1e389dc6f2368409c1b14
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>