Specialize RestconfInvokeOperationsServiceImpl.invokeRpc()
Do not use NormalizedNodePayload, but instead inline handling, so that
we are in explicit control of how input is interpreted.

For XML case we also side-step the need to go through a org.w3c.Document,
speeding up deconding.

This eliminates out the first of the four uses of *NormalizedNodeBodyReader,
but it does not reduce its complexity due to codepaths being shared with
the action invocation (which goes through POST).

JIRA: NETCONF-1128
Change-Id: I2d36b38236dcddc1849133a13c1bc9766f54fc8b
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>