WIP: Specialize RestconfInvokeOperationsServiceImpl.invokeRpc()
Do not use NormalizedNodePayload, but instead inline handling, so that
we are in explicit control of how input is interpreted.

This allows us to explicitly call on DatabindContext's codecs, which
speeding up instantiation. For XML case we also side-step the need to go
through a org.w3c.Document.

// FIXME: codecs are not quite right: they must come from supplied
//        schemacontext (because it can be the mount point)

JIRA: NETCONF-1128
Change-Id: I2d36b38236dcddc1849133a13c1bc9766f54fc8b
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>