Make deviceName not to be Optional#empty
We are passing empty value of deviceName to methods only if the request,
for swagger document, is for controller (not a mounted device).
In that case we want "Controller" in place of the deviceName.

This change replaces the empty value for "Controller" in deviceName,
but still keeps the use of Optional API as method parameter.

JIRA: NETCONF-1089
Change-Id: Idf77e3a42b4a8ecfdc7bef98ec99eeccc151b384
Signed-off-by: Šimon Ukuš <simon.ukus@pantheon.tech>