WIP create test for controller swagger document
This test simulates server and client. This client
then performs calls to the API to which the server
responds accordingly. We then extract the response body
and perform check.

This lets us test the overall result of the calls and verify
contents that are returned back to the client.

JIRA: NETCONF-1131
Change-Id: I008f01e3c11593f5a2d0f53843bc4d5639abda1a
Signed-off-by: Šimon Ukuš <simon.ukus@pantheon.tech>