Fix incorrect schema reference in root POST link
We are creating a root POST link, when requesting swagger document
for single module. We were passing empty string as module name
when building the POST operation, which resulted in our schema
beginning with underscore (like this: "_toaster_module").

We are creating a schema for this, but without the leading underscore.

The call to the buildPost() is altered here, so that the correct
schema is created for this request.

JIRA: NETCONF-1133
Change-Id: I9019dea1874133706f8825b0159e6c571d7b9558
Signed-off-by: Šimon Ukuš <simon.ukus@pantheon.tech>