Make NotificationMessage constructor private
Made NotificationMessage constructor private and added a new public static method wrapDocumentAsNotification,
serving the same purpose. The reason for that is that the wrapping is done in the static method and not the
constructor as before. This enables adding another static method, where the document will not be wrapped, if it
already represents a valid NotificationMessage.

JIRA: NETCONF-1014
Signed-off-by: matus.matok <matus.matok@pantheon.tech>
Change-Id: I74ca0bb09be2652d5179956e5b54c17d9c655ab0