Rework servers record by eliminating JsonNodes
The servers field in the Path record, although not used, is incorrectly
defined as an ObjectNode. The OpenApi Specification declares that such
field is an array of Server Objects.

This change modifies the field declaration accordingly  and extends
the already present Server record by additional fields.
Introducing ServerVariable, conforming to the specification.

JIRA: NETCONF-938
Change-Id: I7b8141c38db18c49585a05e427ef64922db4bf6b
Signed-off-by: Šimon Ukuš <simon.ukus@pantheon.tech>
Signed-off-by: Samuel Schneider <samuel.schneider@pantheon.tech>