Remove AbstractNormalizedNodeBodyReader
We are using this reader only in postData(). Refactor users to
centralized dispatch and move the logic to ChildBody and its
specializations.

This refactor also improves dispatch to datastore resource, as we know
that such POST requests are not action invocations.

Invocation path is using OperationInputBody, so as to share decoding
with the RPC invocation path.

JIRA: NETCONF-1128
Change-Id: Ie632de94dce24d3a194c560ae41e3e86d1e8ac72
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>