Fix incorrect schema reference in root POST link
We are creating a root POST link, when requesting swagger document
for single module. We were passing empty string as module name
when building the POST operation, which resulted in our schema ref
beginning with underscore (like this: "_config_toaster_module").
The schema name was also incorrectly created (missing the "_config").

This patch fixes it by altering the api call to create the POST request
and by fixing the schema name. This also adds the xml module schema,
which was previously missing.

JIRA: NETCONF-1133
Change-Id: I9019dea1874133706f8825b0159e6c571d7b9558
Signed-off-by: Šimon Ukuš <simon.ukus@pantheon.tech>