Optimize AbstractWebsocketSerializer.convertPath()
The serialization has a couple of classic duplications. Eliminate them
through simple addition of appendQName() utility method.

Also drop FIXMEs to examine the output, as it seems extremely fishy:
the namespace is a URN, which means ':' and '-' are not the same thing
and namespaces can very much differ by these.

Also key values are straight toString()'d which means QNames and similar
go through ODL-specific encoding.

Change-Id: I658e02517c1f155ba65ff9ae345228ee015a70aa
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>