WIP: Use NormalizedNodeStreamWriter to emit XML data
We are already instantiating NormalizedNode-based writer, which means we
get help in terms for formatting things. This is especially true of
identityrefs and instance-identifiers, where the XML codec does
absolutely the right thing.

This also allows us to ditch namespace repairing from our
XMLWriterFactory, so we are in explicit control of our namespaces.

// FIXME: convert more test data

JIRA: NETCONF-1152
Change-Id: I030d25a8e0b23a78b6c6e952b36395bd461b4c70
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>