WIP: Refactor AbstractNormalizedNodeBodyWriter
AbstractNormalizedNodeBodyWriter so that the three common dispatches are
handled in common code. This also modernizes them and pulls common bits
together.

This allows us to also switch the dispatch logic, so it operates on
SchemaInferenceStack instead of on SchemaNode.

All of this ends up showing that there is a bug in operation output
code, which causes us to emit an unnecessary declaration.

JIRA: NETCONF-1157
Change-Id: Icc781e884cecf7a5d64de63dd65b97e963fd4abe
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>