WIP: Introduce RequestMode
We are retiring InstanceIdentifierContext in favor of a better
construct.

This construct is RequestMode, which has three specializations:
- ResourceMode, for when the request pertains a particular resource,
  a datastore or in general the global scope
- CreateResourceMode, for when a POST URL indicates we should embark
  on resource creation
- InvokeOperationMode, for when a POST URL indicates we should embark
  on RPC or Action invocation.

The latter are indirect, going through a POSTMode intermediate class,
which allows us to model unified return type for binding PUT methods.

The flow now makes it clear that DatabindProvider should be integrated
into MdsalRestconfServer. For the two cases which do weird things on
local scope, we expose a legacy localDatabind() method.

This further clarifies that ParserIdentifier.toInstanceIdentifier()
cannot be invoked with null identifier, so we get rid of all the
complexity -- eventually guaranteeing that an InstanceIdentifierContext
always has an InstanceIdentifier.

// FIXME: finish conversion

JIRA: NETCONF-1157
Change-Id: I635648d79ad23529d633d43e9c8b1ac1c3497776
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>