Introduce TransportSsh{Client,Server}
We are using both SshClient and SshServer, configuring them extensively,
but we never start them.

Introduce proper specializations of SshClient and SshServer, which do
not allow themselves to be start or stopped. Strictly instantiate them
through provided builders, which also serve as hosts for configuration
adaptation.

This reduces clutter in SSH{Client,Server}, making them single-page
classes. Another benefit is that the consistency of SshClient/SshServer
is now checked, with defaults being correctly applied.

JIRA: NETCONF-590
Change-Id: If68714c21fe06de84cdff743c59c08f0f96f884c
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>