Add SSHTransportStackFactory
TransportStack integration with Apache SSHD requires non-trivial
interaction with Netty, as we need to integrate on both I/O and
threadpool/executor levels.

At the end of the day, anything from SSHD needs to run on worker event
loop (group). This in turn requires having control over the worker
EventLoopGroup.

SSHTransportStackFactory is a resource (e.g. it needs to be
lifecycle-managed), which captures the lifecycle of EventLoopGroups
compatible with NettyTransportSupport, i.e. using the best available
technology stack.

This is the new primary entrypoint into this component, with users
expected to create a SSHTransportStackFactory, manage its lifecycle
and create SSH client/servers from it.

JIRA: NETCONF-590
Change-Id: I80622f1595344a83d439717e243295cb4683383c
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>