Fix filter type namespace
We are emitting "type" with the NETCONF base namespace.

This is wrong, as attributes are not subject to default namespace, hence
we should be emitting it as a plain attribute -- it is interpreted in
the context of its defining element, which already is in NETCONF base
namespace.

JIRA: NETCONF-1170
Change-Id: I9cf63ca52297cd45ea062ec7ddfb1ff4a9e853f7
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>
(cherry picked from commit 1c823018bbdaf99f2ead7cd22fd37d372b4f85b1)