Keep track of connection count in NetconfDevice
Checking the current connection status is not enough to guarantee that
schema setup was completed for the initial session-up event.
Keep track of the connection count, which is incremented every time
onRemoteSessionUp is called, and use the count to make sure that
the session-up event is propagated only when the device stays connected
without reconnection until schema setup is completed. This prevents
unnecessary duplicate notification and avoid confusion in the connection
state.

JIRA: NETCONF-1173
Change-Id: I7f84da4a9197db52a0c33d2640e0c2543e587d7b
Signed-off-by: Sangwook Ha <sangwook.ha@verizon.com>