Keep track of connection count in NetconfDevice
Checking the current connection status is not enough to guarantee that
schema setup was completed before the next session-up event.
Keep track of the connection count, which is incremented every time
onRemoteSessionUp is called and use the count to make sure that
the session-up event is propagated only when schema setup is completed
for the triggering event.

JIRA: NETCONF-1173
Change-Id: I7f84da4a9197db52a0c33d2640e0c2543e587d7b
Signed-off-by: Sangwook Ha <sangwook.ha@verizon.com>