Make ExistenceCheck implement FutureCallback
There is some shared state we are expressing, but we are leaving
implicit dependencies up to an anonymous class to capture.

ExistenceCheck is never instantiated, which seems to be a waste. Split
out the stateful bit allocation to start() and let ExistenceCheck be a
FutureCallback dealing the partial outcome.

Having state ownership indicated a bit more clearly, this exposes a
subtle race: if the last two checks complete concurrently, and one would
result in success and the other in failure, the counter-based logic
could result in a false success.

We document the above interactions and reorder operations to close the
above race without introducing additional locking.

JIRA: NETCONF-718
Change-Id: I3aa8edb4df695131be7e0b79e7f1453fb26837f4
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>