Remove POST request with payload contains only leafs
Remove generation of post operations with payload contains
only leafs or leafrefs.

Current implementation generates the wrong payload for JSON and XML POST requests with leafs or leaf-list in the payload.
With this solution, we can avoid the need to check how many leaves are in the container and what to do if more of them are mandatory.

JIRA: NETCONF-1164, NETCONF-1165
Change-Id: I6268ab118fd0cb5092d4b959660855035394e436
Signed-off-by: Samuel Schneider <samuel.schneider@pantheon.tech>